requirements.txt
icalendar==5.0.1
PyYAML==6.0.1